Moved out the Blosxom-dir provider to it's own bundle. Removed the Doctrine ORM provider. While it's still going to be the default provider, I don't feel like keeping two providers up to date with interface changes at this stage of the project, and the Blosxom-dir provider holds my interest more.
Created a DIC tag and configuration for which provider to use. Added the blosxom data dir as configuration to that bundle.

Abstracted entity interactions to interfaces, because providers may well need to provider their own implementations due to performance concerns.

Added support for entry permalinks, to see just one entry, based on a permalink id.